+++ sfeprapy/mcs/mcs_gen_2.py
@@ -150,7 +150,7 @@ class InputParser(Params2Samples):
 
         for k, v in dist_params.items():
 
-            if isinstance(v, float) or isinstance(v, int) or isinstance(v, np.float):
+            if isinstance(v, float) or isinstance(v, int) or isinstance(v, float):
                 dict_out[k] = np.full((num_samples,), v, dtype=float)
 
             elif isinstance(v, str):
